Dockerized mongoDB and frontend

+++ Backend/clearDB.py
@@ -1,4 +1,7 @@
 import pymongo
 
 client = pymongo.MongoClient("mongodb://localhost:27017/")
-client.drop_database('convrt_database')
+allDBs = client.list_database_names()
+
+if 'convrt_database' in allDBs:
+    client.drop_database('convrt_database')
